// Segundo parcial/Clases/09capast/interfaces/IProducto.ts
// APLICACIONES WEB 2
// Nombres y Apellidos: Winter Aníbal Meza Jiménez.
// Curso: Sexto "B" 2022(1).
// Clase: Arquitectura de capas con Typescript.
// Fecha: Lunes, 06 de mayo de 2022.
// Docente: Ing. Jhon Antonio Cevallos Macías, Mg.
export interface IProducto {
    nombre: String;
    estado: Boolean;
    precio: Number;
    costo: Number;
    minimo:Number;
    stock:Number;
}